Rename InMemoryConnectorRegistry to ConfiguredConnectorRegistry with ImmutableDictionary
Connectors are only registered at startup (bootstrap-time constants) and
read-only thereafter. The old name implied runtime mutability and violated
the "InMemory for dev/test only" architectural constraint. The new name
reflects its actual semantics: a startup-initialized, configuration-driven
registry. The lock is replaced with ImmutableDictionary for lock-free
thread-safe reads. Documentation is added noting that all nodes must be
configured identically in multi-node deployments.
